Read unitprocedures

+++ pkg/fhxModels/procedure.go
@@ -2,8 +2,7 @@ package fhxModels
 
 type Procedure struct {
 	Name       string      `json:"name"`
-	Time       int64       `json:"time"`
+	Time       int         `json:"time"`
 	Author     string      `json:"author"`
 	Parameters []Parameter `json:"params"`
 }
-
